Fix #1 split source
